# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Potoo API
  version: "0.1.0"
  description: |
    Developer-first notification orchestration platform.

    Authenticate all requests with an API key in the `Authorization` header:

    ```
    Authorization: Bearer <api_key>
    ```
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: API key issued from the Potoo dashboard.
  schemas:
    AuthUser:
      type: object
      required: [id, email, created_at]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
    OrgSummary:
      type: object
      required: [id, name, slug, type, role, is_active]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        type:
          type: string
          enum: [platform, customer]
        role:
          type: string
          enum: [owner, member]
        is_active:
          type: boolean
        api_key_count:
          type: integer
          description: Number of active (non-revoked) API keys — included in platform org listing
        call_count:
          type: integer
          description: Total notifications sent — included in platform org listing
    UpdateOrgRequest:
      type: object
      properties:
        is_active:
          type: boolean
    SetupRequest:
      type: object
      required: [org_name, first_name, last_name, email, password]
      properties:
        org_name:
          type: string
          description: Platform organization name
          example: Acme Platform
        org_website:
          type: string
          example: https://acme.com
        org_description:
          type: string
          maxLength: 200
          example: The platform organization
        first_name:
          type: string
          example: James
        last_name:
          type: string
          example: Zaman
        phone:
          type: string
          example: "+447911123456"
        email:
          type: string
          format: email
          example: admin@example.com
        password:
          type: string
          minLength: 8
    SelectOrgRequest:
      type: object
      required: [org_id]
      properties:
        org_id:
          type: string
    CreateOrgRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          example: Acme Corp
        slug:
          type: string
          example: acme-corp
    ApiKey:
      type: object
      required: [id, name, key_prefix, scopes, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        key_prefix:
          type: string
          description: First characters of the key — safe to display, not secret
        scopes:
          type: array
          items:
            type: string
        last_used_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    CreateAPIKeyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: Production key
        scopes:
          type: array
          items:
            type: string
          default: []
    CreateAPIKeyResponse:
      allOf:
        - $ref: "#/components/schemas/ApiKey"
        - type: object
          required: [key]
          properties:
            key:
              type: string
              description: Full API key — shown once, never retrievable again
    RegisterRequest:
      type: object
      required: [org_name, first_name, last_name, email, password]
      properties:
        org_name:
          type: string
          example: Skyline Labs
        org_website:
          type: string
          example: https://example.com
        org_description:
          type: string
          maxLength: 200
          example: A short description of what your organization does
        first_name:
          type: string
          example: Morgan
        last_name:
          type: string
          example: Taylor
        phone:
          type: string
          example: "+1 555 000 0000"
        email:
          type: string
          format: email
          example: you@example.com
        password:
          type: string
          minLength: 8
    InviteAcceptRequest:
      type: object
      required: [token, first_name, last_name, email, password]
      properties:
        token:
          type: string
        first_name:
          type: string
          example: James
        last_name:
          type: string
          example: Zaman
        phone:
          type: string
          example: "+447911123456"
        email:
          type: string
          format: email
          example: james@example.com
        password:
          type: string
          minLength: 8
    CreateInviteRequest:
      type: object
      required: [org_id]
      properties:
        org_id:
          type: string
    InviteResponse:
      type: object
      required: [id, org_id, token, expires_at, created_at]
      properties:
        id:
          type: string
        org_id:
          type: string
        org_name:
          type: string
        token:
          type: string
        used_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    InvitePreview:
      type: object
      required: [org_name, expires_at]
      properties:
        org_name:
          type: string
        expires_at:
          type: string
          format: date-time
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              example: validation_failed
            message:
              type: string
              example: "Template is missing required variable: first_name"
            request_id:
              type: string
              example: req_01HX1234
            docs_url:
              type: string
              example: https://docs.potoo.com/errors/validation_failed
...
security:
  - ApiKeyAuth: []
paths:
  /v1/auth/register:
    post:
      operationId: register
      summary: Create a customer organization and owner account together
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Organization and account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error or platform not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Email or org slug already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/auth/accept-invite:
    post:
      operationId: acceptInvite
      summary: Accept an invite and create a user account
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteAcceptRequest"
      responses:
        "201":
          description: Account created and joined organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: Invite already used or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/invite/{token}:
    get:
      operationId: getInvite
      summary: Validate an invite token and return org info for the invite page
      tags: [auth]
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Valid invite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitePreview"
        "404":
          description: Invite not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: Invite already used or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/auth/login:
    post:
      operationId: login
      summary: Sign in and receive a session cookie
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated — sets HttpOnly session cookie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/auth/refresh:
    post:
      operationId: refreshSession
      summary: Exchange a refresh token for a new access + refresh token pair
      tags: [auth]
      security: []
      responses:
        "200":
          description: New tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/auth/me:
    get:
      operationId: getMe
      summary: Get the current authenticated user
      tags: [auth]
      security: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/setup:
    post:
      operationId: setup
      summary: One-time platform bootstrap — creates platform org and first owner
      tags: [setup]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupRequest"
      responses:
        "201":
          description: Platform bootstrapped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: Platform already configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/auth/select-org:
    post:
      operationId: selectOrg
      summary: Switch active org — deletes current session and starts a new one
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectOrgRequest"
      responses:
        "200":
          description: Org selected — new session issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not a member of the requested org
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/me/orgs:
    get:
      operationId: listMyOrgs
      summary: List all orgs the current user belongs to
      tags: [auth]
      security: []
      responses:
        "200":
          description: Org list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/OrgSummary"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/platform/orgs:
    get:
      operationId: listOrgs
      summary: List all customer organizations (platform owner only)
      tags: [platform]
      responses:
        "200":
          description: Org list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/OrgSummary"
    post:
      operationId: createOrg
      summary: Create a customer organization (platform owner only)
      tags: [platform]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrgRequest"
      responses:
        "201":
          description: Org created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgSummary"
        "409":
          description: Slug already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/platform/orgs/{orgId}:
    patch:
      operationId: updateOrg
      summary: Update a customer organization (platform owner only)
      tags: [platform]
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrgRequest"
      responses:
        "200":
          description: Org updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgSummary"
        "404":
          description: Org not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/platform/invites:
    post:
      operationId: createInvite
      summary: Create a single-use invite link for a customer org (platform owner only)
      tags: [platform]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviteRequest"
      responses:
        "201":
          description: Invite created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteResponse"
        "404":
          description: Org not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listInvites
      summary: List all invites (platform owner only)
      tags: [platform]
      security: []
      responses:
        "200":
          description: Invite list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/InviteResponse"
  /v1/api-keys:
    get:
      operationId: listAPIKeys
      summary: List API keys for the session org
      tags: [api-keys]
      security: []
      responses:
        "200":
          description: Key list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApiKey"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createAPIKey
      summary: Create an API key
      tags: [api-keys]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAPIKeyRequest"
      responses:
        "201":
          description: Key created — full key shown once
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAPIKeyResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not authorised to create a key for that org
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/api-keys/{keyId}:
    delete:
      operationId: revokeAPIKey
      summary: Revoke an API key
      tags: [api-keys]
      security: []
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Key revoked
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/webhooks/email/{provider}:
    post:
      operationId: ingestEmailWebhook
      summary: Receive email provider webhook
      description: |
        Receives delivery status events from an email provider (Resend, SendGrid).
        This endpoint is public — no API key required. Signature verification is
        performed using the webhook key stored on the provider connection.
      tags: [webhooks]
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [resend, sendgrid]
          example: resend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Webhook accepted
        "400":
          description: Invalid payload or signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"